Hide senha Psicologo

+++ src/controller/pacienteController.js
@@ -1,60 +1,69 @@
 const { Paciente, Psicologos } = require('../model/');
+const authController = require('../controller/authController');
 
 const pacienteController = {
 
 
 
-    // async buscarPacientes(req, res) {
-    //     try {
-
-    //         const pacientes = await Paciente.findAll({
-    //             include: Psicologos,
-
-    //         });
-
-    //         return res.status(200).json(pacientes);
-    //     } catch (error) {
-    //         return res.status(404).json('Falha ao buscar lista de pacientes');
-    //     }
-    // },
-
-    async paginacaoBuscarPacientes(req, res) {
+    async buscarPacientes(req, res) {
         try {
-            const { termo, page = 1 , limit = 10 } = req.query;
-            const offSet = parseInt(limit) * (parseInt(page) - 1);
 
-            let filter = {
-                limit: parseInt(limit),
-                offSet,
-                include: Psicologos,
-
-
-            }
-
-            // if (termo) {
-            //     Object.assign(filter, {
-            //         where: {
-            //             // person_name: { [Op.like]: `%${termo}%´ }  // Equivalente ao código da linha 11
-            //             nome: { [Op.substring]: termo }
-            //         },
-            //     })
-            // }
+            const pacientes = await Paciente.findAll({
+                include: {
+                    model: Psicologos,
+                    attributes: ['id', 'nome', 'email', 'apresentacao'],
+                },
 
 
 
+            });
 
-            const paginacaoPacientes = await Paciente.findAll(filter)
-            return res.json(paginacaoPacientes);
+            return res.status(200).json(pacientes);
         } catch (error) {
-
+            return res.status(404).json('Falha ao buscar lista de pacientes');
         }
     },
+
+    // async paginacaoBuscarPacientes(req, res) {
+    //     try {
+    //         const { termo, page , limit  } = req.query;
+    //         const offSet = parseInt(limit) * (parseInt(page) - 1);
+
+    //         let filter = {
+    //             limit: parseInt(limit),
+    //             offSet,
+    //             include: Psicologos,
+
+
+    //         }
+
+    //         // if (termo) {
+    //         //     Object.assign(filter, {
+    //         //         where: {
+    //         //             // person_name: { [Op.like]: `%${termo}%´ }  // Equivalente ao código da linha 11
+    //         //             nome: { [Op.substring]: termo }
+    //         //         },
+    //         //     })
+    //         // }
+
+
+
+
+    //         const paginacaoPacientes = await Paciente.findAll(filter)
+    //         return res.json(paginacaoPacientes);
+    //     } catch (error) {
+
+    //     }
+    // },
 
     async buscarPacienteId(req, res) {
         try {
             const { id } = req.params;
             const pacienteId = await Paciente.findByPk(id, {
-                include: Psicologos,
+                include: {
+                    model: Psicologos,
+                    attributes: ['id', 'nome', 'email', 'apresentacao'],
+                },
             });
             if (pacienteId == null) {
                 return res.status(404).json('Id não encontrado');
@@ -75,17 +84,18 @@ const pacienteController = {
     async cadastrarPaciente(req, res) {
         try {
             const { nome, email, idade, psicologos_id } = req.body;
-
+          
+            
             const novoPaciente = await Paciente.create({
                 nome,
                 email,
                 idade,
-                psicologos_id
+                psicologos_id 
             });
-           
-           
+            
             const psicologos = await Psicologos.findByPk(psicologos_id);
             await novoPaciente.setPsicologos(psicologos);
+            
             
 
             return res.status(201).json(novoPaciente);
@@ -127,6 +137,11 @@ const pacienteController = {
     async deletarPaciente(req, res) {
         try {
             const { id } = req.params;
+            const pacienteId = await Paciente.findByPk(id);
+
+            if (pacienteId == null) {
+                return res.status(404).json('Id não encontrado');
+            };
 
             await Paciente.destroy({
                 where: {
